i deleted the logging thing because i dont nned it rn
